♻️(web) regorganize presentation tests for ats
